% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - fishtools.R
\name{calc_growth_TGC}
\alias{calc_growth_TGC}
\title{predicts growth based on TGC, same model as above}
\usage{
calc_growth_TGC(W0, TGC, temperature, time)
}
\description{
predicts growth based on TGC, same model as above
}
