% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - fishtools.R
\name{calc_TGC}
\alias{calc_TGC}
\title{calculates thermal growth coefficient (from Jobling: The thermal growth coefficient (TGC) model of fish growth: a cautionary note)}
\usage{
calc_TGC(W0, W1, temperature, time)
}
\description{
calculates thermal growth coefficient (from Jobling: The thermal growth coefficient (TGC) model of fish growth: a cautionary note)
}
