% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - fishtools.R
\name{calc_TGC_df}
\alias{calc_TGC_df}
\title{For a whole dataframe, calculates Thermal Growth Coefficient TGC for a given period}
\usage{
calc_TGC_df(tb_fish, tb_temp, period)
}
\description{
For a whole dataframe, calculates Thermal Growth Coefficient TGC for a given period
}
