% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R
\name{calculate_distances}
\alias{calculate_distances}
\title{Calculate Distances Between Spots In Opposite Channels}
\usage{
calculate_distances(ch1_spots, ch2_spots)
}
\arguments{
\item{ch1_spots}{Data frame containing spot data from first channel}

\item{ch2_spots}{Data frame containing spot data from second channel}
}
\value{
Data frame containing all pairwise distances between spots in opposite channels
}
\description{
Calculate Distances Between Spots In Opposite Channels
}
