% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_neighbours.R
\name{closest_neighbours}
\alias{closest_neighbours}
\title{Identify Closest Neighbours In Opposite Channels}
\usage{
closest_neighbours(distances)
}
\arguments{
\item{distances}{Data frame containing all pairwise distances between spots in opposite channels}
}
\value{
List of two data frames containing closest neighbours for spots in each channel
}
\description{
Identify Closest Neighbours In Opposite Channels
}
