% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_threshold_test.R
\name{coloc_threshold_test}
\alias{coloc_threshold_test}
\title{Calculate Colocalisation At A Range Of Distance Thresholds}
\usage{
coloc_threshold_test(
  spots,
  neighbours,
  thresholds = seq(0, 1000, 50),
  iterations = 5
)
}
\arguments{
\item{spots}{List containing cell and spot data (from which closest neighbours have been calculated)}

\item{neighbours}{Data frame containing closest neighbours for each spot}

\item{thresholds}{Numeric vector containing the distance thresholds to test}

\item{iterations}{The number of iterations to perform for the null distributions}
}
\value{
Data frame containing colocalisation percentage at each threshold
}
\description{
Calculate Colocalisation At A Range Of Distance Thresholds
}
