% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalisation_test.R
\name{colocalisation_test}
\alias{colocalisation_test}
\title{Use Threshold To Decide Whether Spots Are Colocalised}
\usage{
colocalisation_test(
  distances,
  method = "absolute",
  threshold = 500,
  multiplier = 1
)
}
\arguments{
\item{distances}{List of two data frames containing distances between nearest pairs of spots}

\item{method}{Method used to determine colocalisation. Options are "radius" (uses sum of spot radii) or "absolute" (sets an absolute distance)}

\item{threshold}{Distance threshold (in nm) for determining colocalisation. Used if method = "absolute"}

\item{multiplier}{Multiplier for sum of radii. Set lower for more strict colocalisation. Used if method = "radius"}
}
\value{
List of two data frames with added columns indicating whether each pair of spots is colocalised or not (TRUE or FALSE)
}
\description{
Use Threshold To Decide Whether Spots Are Colocalised
}
