% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances}
\alias{distances}
\title{Calculate Distances Between Spots In Opposite Channels For A Single Cell}
\usage{
distances(ch1_spots, ch2_spots, cell_of_interest)
}
\arguments{
\item{ch1_spots}{Data frame containing spot data from first channel}

\item{ch2_spots}{Data frame containing spot data from second channel}

\item{cell_of_interest}{The index of the cell of interest}
}
\value{
Data frame containing all pairwise distances between spots in opposite channels
}
\description{
Calculate Distances Between Spots In Opposite Channels For A Single Cell
}
