% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - fishtools.R
\name{extract_periods}
\alias{extract_periods}
\title{Takes a wide-format dataset of individuals, with period-associated measurements denoted by dot-number format, e.g:
Weight.1, weight.2, weight.3 or length.1 length.2 or date.1, date2. etc
Returns a numeric vector of all the periods in that dataset eg: c(1,2,3,4,5) for a dataset with 5 periods}
\usage{
extract_periods(tb_fish)
}
\description{
Takes a wide-format dataset of individuals, with period-associated measurements denoted by dot-number format, e.g:
Weight.1, weight.2, weight.3 or length.1 length.2 or date.1, date2. etc
Returns a numeric vector of all the periods in that dataset eg: c(1,2,3,4,5) for a dataset with 5 periods
}
