% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intersection.R
\name{find_intersection}
\alias{find_intersection}
\title{Estimate The Intersection Between Two Gaussian Mixture Model Components}
\usage{
find_intersection(
  spots,
  model,
  component_1 = NULL,
  component_2 = NULL,
  n_bins = 120
)
}
\arguments{
\item{spots}{Data frame containing spot data}

\item{model}{Gaussian mixture model parameters, the output of fish_mclust}

\item{component_1}{The first component of the model to compare. The first component must have the LOWER mean.}

\item{component_2}{The second component pof the model to compare.}

\item{n_bins}{Number of bins to use (default = 120)}
}
\value{
A single value corresponding to the estimated intersection point (raw intensity) between the two components of interest
}
\description{
Estimate The Intersection Between Two Gaussian Mixture Model Components
}
