% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_cell_outlines.R
\name{fish_cell_outlines}
\alias{fish_cell_outlines}
\title{Extract Cell Outlines From FISH-QUANT File}
\usage{
fish_cell_outlines(file, line_selection)
}
\arguments{
\item{file}{File from FISH-QUANT (imported using readLines)}

\item{line_selection}{Data frame containing the line indices for each cell}
}
\value{
Data frame containing the cell outlines in tidy format
}
\description{
Extract Cell Outlines From FISH-QUANT File
}
