% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_closest_spots.R
\name{fish_closest_spots}
\alias{fish_closest_spots}
\title{Identify Closest Neighbours In Opposite Channels}
\usage{
fish_closest_spots(distances)
}
\arguments{
\item{distances}{Data frame containing all pairwise distances between spots in opposite channels}
}
\value{
List of two data frames containing closest neighbours for spots in each channel
}
\description{
Identify Closest Neighbours In Opposite Channels
}
