% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_coloc_threshold.R
\name{fish_coloc_threshold}
\alias{fish_coloc_threshold}
\title{Use Threshold To Decide Whether Spots Are Colocalised}
\usage{
fish_coloc_threshold(
  distances,
  method = "absolute",
  threshold = 500,
  multiplier = 1
)
}
\arguments{
\item{distances}{Data frame containing distances between pairs of spots}

\item{method}{Method used to determine colocalisation. Options are "radius" (uses sum of spot radii) or "absolute" (sets an absolute distance)}

\item{threshold}{Distance threshold (in nm) for determining colocalisation. Used if method = "absolute"}

\item{multiplier}{Multiplier for sum of radii. Set lower for more strict colocalisation. Used if method = "radius"}
}
\value{
Data frame with an added column indicating whether each pair of spots is colocalised or not (TRUE or FALSE)
}
\description{
Use Threshold To Decide Whether Spots Are Colocalised
}
