% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_colocalisation.R
\name{fish_colocalisation}
\alias{fish_colocalisation}
\title{Calculate Distances Between Spots In Opposite Channels}
\usage{
fish_colocalisation(spots, method = "radius", multiplier = 1, threshold = 500)
}
\arguments{
\item{spots}{Data frame containing spot data in two channels}

\item{method}{Method used to determine colocalisation. Options are "radius" (uses sum of spot radii) or "absolute" (sets an absolute distance)}

\item{multiplier}{Multiplier for sum of radii. Set lower for more strict colocalisation. Used if method = "radius"}

\item{threshold}{Distance threshold (in nm) for determining colocalisation. Used if method = "absolute"}
}
\value{
Data frame containing all pairwise distances between spots in opposite channels
}
\description{
Calculate Distances Between Spots In Opposite Channels
}
