% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_combined.R
\name{fish_combined}
\alias{fish_combined}
\title{Extract Cell Outlines And Spot Data From FISH-QUANT File}
\usage{
fish_combined(path, channel)
}
\arguments{
\item{path}{The file path of the FISH-QUANT output}

\item{channel}{The channel being analysed e.g. "Cy5"}
}
\value{
List containing data frames of cell outlines and spot data
}
\description{
Extract Cell Outlines And Spot Data From FISH-QUANT File
}
