% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_import.R
\name{fish_import}
\alias{fish_import}
\title{Extract Spot Data From FISH-QUANT File}
\usage{
fish_import(path, channel = NA, mRNA = NA)
}
\arguments{
\item{path}{The file path of the FISH-QUANT output}

\item{channel}{The channel being analysed e.g. "Cy5"}

\item{mRNA}{The name of the mRNA being analysed e.g. "PGK1"}
}
\value{
Data frame containing spot data
}
\description{
Extract Spot Data From FISH-QUANT File
}
