% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_line_selection.R
\name{fish_line_selection}
\alias{fish_line_selection}
\title{Determine Which Lines To Use From FISH-QUANT File}
\usage{
fish_line_selection(file)
}
\arguments{
\item{file}{File from FISH-QUANT (imported using readLines)}
}
\value{
Data frame containing the line indices for each cell
}
\description{
Determine Which Lines To Use From FISH-QUANT File
}
