% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_mclust.R
\name{fish_mclust}
\alias{fish_mclust}
\title{Fit A Gaussian Mixture Model To FISH Spot Data}
\usage{
fish_mclust(data)
}
\arguments{
\item{data}{Output list from fish_combined, containing cell outlines and spot data}
}
\value{
Parameters for a Gaussian mixture model from mclust
}
\description{
Fit A Gaussian Mixture Model To FISH Spot Data
}
