% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_null_distribution.R
\name{fish_null_distribution}
\alias{fish_null_distribution}
\title{Estimate Null Distribution For Colocalisation Data}
\usage{
fish_null_distribution(spots, iterations = 10)
}
\arguments{
\item{spots}{List containing spot data}

\item{iterations}{Number of iterations to perform, default = 10}
}
\value{
Data frame containing simulated \% colocalisation for each channel (one row per iteration)
}
\description{
Estimate Null Distribution For Colocalisation Data
}
