% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_plot_all.R
\name{fish_plot_all}
\alias{fish_plot_all}
\title{Plot FISH-QUANT Data For All Cells}
\usage{
fish_plot_all(data)
}
\arguments{
\item{data}{Output list from fish_combined, containing cell outlines and spot data}
}
\value{
ggplot object showing all cells and spots
}
\description{
Plot FISH-QUANT Data For All Cells
}
