% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_plot_circle.R
\name{fish_plot_circle}
\alias{fish_plot_circle}
\title{Plot FISH-QUANT Data For A Single Cell, With Spots Proportionately Sized}
\usage{
fish_plot_circle(data, cell_of_interest)
}
\arguments{
\item{data}{Output list from fish_combined, containing cell outlines and spot data}

\item{cell_of_interest}{The index of the cell to plot}
}
\value{
ggplot object showing cell of interest with all of its spots, with spots proportionately sized
}
\description{
Plot FISH-QUANT Data For A Single Cell, With Spots Proportionately Sized
}
