% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_plot_threshold.R
\name{fish_plot_threshold}
\alias{fish_plot_threshold}
\title{Plot FISH-QUANT Data For A Single Cell Along With Colocalisation Threshold}
\usage{
fish_plot_threshold(data, cell_of_interest, threshold = 300)
}
\arguments{
\item{data}{Output list from fish_combined, containing cell outlines and spot data}

\item{cell_of_interest}{The index of the cell to plot}

\item{threshold}{Distance threshold to plot (in nm)}
}
\value{
ggplot object showing cell of interest with all of its spots, with spots proportionately sized
}
\description{
Plot FISH-QUANT Data For A Single Cell Along With Colocalisation Threshold
}
