% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_outlines.R
\name{import_outlines}
\alias{import_outlines}
\title{Extract Cell Outlines From FISH-QUANT File}
\usage{
import_outlines(path)
}
\arguments{
\item{path}{The file path of the FISH-QUANT output}
}
\value{
Data frame containing cell outlines
}
\description{
Extract Cell Outlines From FISH-QUANT File
}
