% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_selection.R
\name{line_selection}
\alias{line_selection}
\title{Determine Which Lines To Use From FISH-QUANT File}
\usage{
line_selection(file)
}
\arguments{
\item{file}{File from FISH-QUANT (imported using readLines)}
}
\value{
Data frame containing the line indices for each cell
}
\description{
Determine Which Lines To Use From FISH-QUANT File
}
