% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust_plot_mRNA.R
\name{mclust_plot_mRNA}
\alias{mclust_plot_mRNA}
\title{Plot Histogram And Gaussian Mixture Model From Transformed Spot Data}
\usage{
mclust_plot_mRNA(data, model, n_bins = 120)
}
\arguments{
\item{data}{Output list from transform_data, containing number_of_mRNAs column in spots data frame}

\item{model}{Transformed Gaussian mixture model parameters to use. The output of fish_mclust and transform_model.}

\item{n_bins}{Number of bins to use (default = 120)}
}
\value{
A histogram with overlaid Gaussian mixture model
}
\description{
Plot Histogram And Gaussian Mixture Model From Transformed Spot Data
}
