% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust_spots.R
\name{mclust_spots}
\alias{mclust_spots}
\title{Fit A Gaussian Mixture Model To FISH Spot Data}
\usage{
mclust_spots(spots)
}
\arguments{
\item{spots}{Data frame containing spot data}
}
\value{
Parameters for a Gaussian mixture model from mclust
}
\description{
Fit A Gaussian Mixture Model To FISH Spot Data
}
