% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_shuffle_spots.R
\name{null_shuffle_spots}
\alias{null_shuffle_spots}
\title{Shuffle Channel Two Spots And Calculate Colocalisation}
\usage{
null_shuffle_spots(
  ch1_spots,
  ch2_spots,
  method = "absolute",
  threshold = 500,
  multiplier = 1
)
}
\arguments{
\item{ch1_spots}{Data frame containing spot data from first channel}

\item{ch2_spots}{Data frame containing spot data from second channel}

\item{method}{Method used to determine colocalisation. Options are "radius" (uses sum of spot radii) or "absolute" (sets an absolute distance)}

\item{threshold}{Distance threshold (in nm) for determining colocalisation. Used if method = "absolute"}

\item{multiplier}{Multiplier for sum of radii. Set lower for more strict colocalisation. Used if method = "radius"}
}
\value{
List of two data frames indicating whether spots are colocalised
}
\description{
Shuffle Channel Two Spots And Calculate Colocalisation
}
