% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_funs.R
\name{OK}
\alias{OK}
\title{Error catching for try}
\usage{
OK(object)
}
\arguments{
\item{object}{evaluation of the try function}
}
\description{
Error catchinig for use with the try function
}
\details{
Written by Bill Venables
}

