% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spots.R
\name{plot_spots}
\alias{plot_spots}
\title{Plot FISH-QUANT Data For A Single Cell}
\usage{
plot_spots(spots, outlines, cell_of_interest)
}
\arguments{
\item{spots}{Data frame containing spot data}

\item{outlines}{Data frame containing cell outlines}

\item{cell_of_interest}{The index of the cell to plot}
}
\value{
ggplot object showing cell of interest with all of its spots
}
\description{
Plot FISH-QUANT Data For A Single Cell
}
