% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spots_all.R
\name{plot_spots_all}
\alias{plot_spots_all}
\title{Plot FISH-QUANT Data For All Cells}
\usage{
plot_spots_all(spots, outlines)
}
\arguments{
\item{spots}{Data frame containing spot data}

\item{outlines}{Data frame containing cell outlines}
}
\value{
ggplot object showing all cells and spots
}
\description{
Plot FISH-QUANT Data For All Cells
}
