% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spots_circle.R
\name{plot_spots_circle}
\alias{plot_spots_circle}
\title{Plot FISH-QUANT Data For A Single Cell, With Spots Proportionately Sized}
\usage{
plot_spots_circle(spots_1 = NULL, spots_2 = NULL, outlines, cell_of_interest)
}
\arguments{
\item{spots_1}{Data frame containing spot data}

\item{spots_2}{Data frame containing spot data}

\item{outlines}{Data frame containing cell outlines}

\item{cell_of_interest}{The index of the cell to plot}
}
\value{
ggplot object showing cell of interest with all of its spots, with spots proportionately sized
}
\description{
Plot FISH-QUANT Data For A Single Cell, With Spots Proportionately Sized
}
