% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - fishtools.R
\name{predict_weights}
\alias{predict_weights}
\title{predict_weights}
\usage{
predict_weights(
  tb_fish,
  tb_temp,
  dates,
  TGC_baseline = NULL,
  period_start = NULL,
  tgc_p_modifier = 1,
  temp_abs_modifier = 0
)
}
\arguments{
\item{tb_fish}{The table of fish growth data to predict weight for. Table must be in "wide" format; that is, one row pr individual, and columns weight.1, weight.2, weight.3 ... date.1, date.2, date.3... etc for measured weights and their corresponding dates for different periods. Weights must be numerical. Dates must be in date format. The table can contain other columns, these will be ignored. Periods end when the fish has been weighed, so period 2, for example, are the dates between date.1 and date.2. Weight.2 is thus the weight at the end of period 2.}

\item{tb_temp}{A table that contains temperature data for the fish in tb_fish. Needs two columns: date and temp. Column date must be in date format. Is used to calculate mean temperatures between periods.}

\item{dates}{A vector of dates to predict new weights for. Must be contained in a vector and formatted as dates (see example)}

\item{TGC_baseline}{If fish have only been measured once (there's just one period in the table), a thermal growth coefficient can't be calculated, so a TGC needs to be assumed (check data for your species).}

\item{period_start}{optional, if for example set to 3, then periods 1 and 2 will be ignored when estimating thermal growth coefficients.}

\item{tgc_p_modifier}{optional, allows the thermal growth coefficients to be modified by. If tgc_p_modifier is for example set to 1.2, then all thermal growth coefficients are multiplied by 1.2}

\item{temp_abs_modifier}{optional, allows future temperatures to be modified. Setting this to 2 will increase predicted temperature by 2.}
}
\description{
applies to a whole table of weight data for fish, combines growth data with temperature data to predict future weights. Uses Jobling's TGC formula.
}
\examples{
tb_fish_predicted <- predict_weight(tb_fish,tb_temp,c(ymd(20210601),ymd(20211001),ymd(20211201)))
tb_fish_predicted <- predict_weight(tb_fish,tb_temp,today()+weeks(1:5))
}
