% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_funs.R
\name{resample}
\alias{resample}
\title{Sampling convenience function}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{object to sample}

\item{...}{additional parameters for sample.int}
}
\description{
Sampling convenience function for bootstrapping
}
\details{
Example in help for sample function
}

