% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singles_multis_extract.R
\name{singles_multis_extract}
\alias{singles_multis_extract}
\title{Separate Single and Multi mRNA Foci}
\usage{
singles_multis_extract(spots, threshold = 1.5)
}
\arguments{
\item{spots}{Data frame containing spot data, converted to mRNAs per cell}

\item{threshold}{Threshold above which a spot is considered a multi mRNA foci}
}
\value{
List containing data frames of single and multi mRNA foci
}
\description{
Separate Single and Multi mRNA Foci
}
