% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singles_multis_plot.R
\name{singles_multis_plot}
\alias{singles_multis_plot}
\title{Plot Percentage of mRNAs in Multi-mRNA Foci}
\usage{
singles_multis_plot(data)
}
\arguments{
\item{data}{Data frame containing percentage of mRNA copies that are in multi-mRNA foci, the output from summarise_singles_multis}
}
\value{
Plot of percentage of mRNAs in multi-mRNA foci, by mRNA
}
\description{
Plot Percentage of mRNAs in Multi-mRNA Foci
}
