% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_extraction.R
\name{spot_extraction}
\alias{spot_extraction}
\title{Extract Spot Data For A Single Cell From FISH-QUANT File}
\usage{
spot_extraction(file, start, end)
}
\arguments{
\item{file}{File from FISH-QUANT (imported using readLines)}

\item{start}{Line index to start reading from ("SPOTS_START")}

\item{end}{Line index to finish reading from ("SPOTS_END")}
}
\value{
Data frame containing spot data for a single cell in tidy format
}
\description{
Extract Spot Data For A Single Cell From FISH-QUANT File
}
