% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_outlines.R
\name{standardise_outlines}
\alias{standardise_outlines}
\title{Standardise Cell Outline Position Data From FISH-QUANT}
\usage{
standardise_outlines(spots, outlines, pixel_size = NULL)
}
\arguments{
\item{spots}{Data frame containing spot data}

\item{outlines}{Data frame containing cell outlines}

\item{pixel_size}{The size of each pixel (in nanometers). If NULL (default), calculated from spot data: ratio of spot position in nm to spot position in pixels}
}
\value{
Data frame containing cell outlines, with standardised positions
}
\description{
Standardise Cell Outline Position Data From FISH-QUANT
}
