% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_spots.R
\name{standardise_spots}
\alias{standardise_spots}
\title{Standardise Spot Position Data From FISH-QUANT}
\usage{
standardise_spots(spots, outlines, pixel_size = NULL)
}
\arguments{
\item{spots}{Data frame containing spot data}

\item{outlines}{Data frame containing cell outlines}

\item{pixel_size}{The size of each pixel (in nanometers). If NULL (default), calculated from spot data: ratio of spot position in nm to spot position in pixels}
}
\value{
Data frame containing spot data, with standardised positions
}
\description{
Standardise Spot Position Data From FISH-QUANT
}
