% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - fishtools.R
\name{tempData_mean}
\alias{tempData_mean}
\title{Function for getting the mean temperature of a time period
Dates must be supplied as a date object}
\usage{
tempData_mean(tb_temp, date_start, date_end, column_temp)
}
\description{
Function for getting the mean temperature of a time period
Dates must be supplied as a date object
}
