% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform Spot Data To Number Of mRNAs Per Spot}
\usage{
transform_data(spots, model, threshold = 0, single_mRNA_component)
}
\arguments{
\item{spots}{Data frame containing spot data}

\item{model}{Gaussian mixture model parameters, the output of fish_mclust}

\item{threshold}{Raw intensity threshold to use for filtering spot data}

\item{single_mRNA_component}{The component of the model that represents single mRNA spots}
}
\value{
A transformed version of the data containing the column number_of_mRNAs, which is a transformation of the raw intensity
}
\description{
Transform Spot Data To Number Of mRNAs Per Spot
}
