% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_model.R
\name{transform_model}
\alias{transform_model}
\title{Transform Model Parameters To Number Of mRNAs Per Spot}
\usage{
transform_model(model, single_mRNA_component)
}
\arguments{
\item{model}{Gaussian mixture model parameters, the output of fish_mclust}

\item{single_mRNA_component}{The component of the model that represents single mRNA spots}
}
\value{
A transformed version of the model containing the column number_of_mRNAs
}
\description{
Transform Model Parameters To Number Of mRNAs Per Spot
}
