% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances}
\alias{distances}
\title{Calculate Distances Between Spots In Opposite Channels For A Single Cell}
\usage{
distances(spots, cell_of_interest)
}
\arguments{
\item{spots}{Data frame containing spot data in two channels}

\item{cell_of_interest}{The index of the cell of interest}
}
\value{
Data frame containing all pairwise distances between spots in opposite channels
}
\description{
Calculate Distances Between Spots In Opposite Channels For A Single Cell
}
