% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_colocalisation.R
\name{fish_colocalisation}
\alias{fish_colocalisation}
\title{Calculate Distances Between Spots In Opposite Channels}
\usage{
fish_colocalisation(spots)
}
\arguments{
\item{spots}{Data frame containing spot data in two channels}
}
\value{
Data frame containing all pairwise distances between spots in opposite channels
}
\description{
Calculate Distances Between Spots In Opposite Channels
}
