% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust_plot.R
\name{mclust_plot}
\alias{mclust_plot}
\title{Plot Histogram And Gaussian Mixture Model From Spot Data}
\usage{
mclust_plot(spots, model, n_bins = 120)
}
\arguments{
\item{spots}{Data frame containing spot data}

\item{model}{Gaussian mixture model parameters to use. The output of fish_mclust.}

\item{n_bins}{Number of bins to use (default = 120)}
}
\value{
A histogram with overlaid Gaussian mixture model
}
\description{
Plot Histogram And Gaussian Mixture Model From Spot Data
}
