% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singles_multis_summarise.R
\name{singles_multis_summarise}
\alias{singles_multis_summarise}
\title{Summarise Number of Singles and Multis Per Cell in Spot Data}
\usage{
singles_multis_summarise(spots)
}
\arguments{
\item{spots}{Data frame containing spot data, converted to mRNAs per cell}
}
\value{
Data frame containing summary data for sample of interest
}
\description{
Summarise Number of Singles and Multis Per Cell in Spot Data
}
